+++ app/uploaders/workers_uploader.rb
@@ -1,25 +1,2 @@
-class WorkersUploader < CarrierWave::Uploader::Base
-  # Include RMagick or MiniMagick support:
-  include CarrierWave::RMagick
-  # include CarrierWave::MiniMagick
-
-  # 画像の上限を100pxにする
-  process resize_to_limit: [100, 100]
-
-  # Choose what kind of storage to use for this uploader:
-  if Rails.env.development?
-    storage :file
-  elsif Rails.env.test?
-    storage :file
-  else
-    storage :fog
-  end
-
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
-
-  def extension_allowlist
-    %w[jpg jpeg png]
-  end
+class WorkersUploader < Base
 end
