+++ config/initializers/carrierwave.rb
@@ -3,20 +3,21 @@ require 'carrierwave/storage/file'
 require 'carrierwave/storage/fog'
 
 CarrierWave.configure do |config|
-  if Rails.env.production? # 本番環境の場合はS3へアップロード
-    config.storage :fog
-    config.fog_provider = 'fog/aws'
-    config.fog_directory = ENV['S3_BUCKET_NAME'] # バケット名
-    config.fog_public = false
-    config.fog_credentials = {
-      provider:              'AWS',
-      aws_access_key_id:     ENV['S3_ACCESS_KEY_ID'],
-      aws_secret_access_key: ENV['S3_SECRET_ACCESS_KEY'],
-      region:                'ap-northeast-1',
-      path_style:            true
-    }
-  else # 本番環境以外の場合はアプリケーション内にアップロード
-    config.storage :file
-    config.enable_processing = false if Rails.env.test?
-  end
+  # if Rails.env.production? # 本番環境の場合はS3へアップロード
+  config.storage :fog
+  config.fog_provider = 'fog/aws'
+  config.fog_directory = ENV['S3_BUCKET_NAME'] # バケット名
+  config.asset_host = ENV['S3_ASSET_HOST']
+  config.fog_public = false
+  config.fog_credentials = {
+    provider:              'AWS',
+    aws_access_key_id:     ENV['S3_ACCESS_KEY_ID'],
+    aws_secret_access_key: ENV['S3_SECRET_ACCESS_KEY'],
+    region:                'ap-northeast-1',
+    path_style:            true
+  }
+  # else # 本番環境以外の場合はアプリケーション内にアップロード
+  #   config.storage :file
+  #   config.enable_processing = false if Rails.env.test?
+  # end
 end
